requirements.txt
python-dotenv
aiogram
sqlalchemy
gino
psycopg2
datetime
googlemaps